Shared:Components:RoomType: move from client and rewrite to typescript
