+++ packages/client/src/components/dialogs/CreateEditRoomDialog/sub-components/RoomTypeDropdown/DropdownDesktop.js
@@ -27,8 +27,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import RoomType from "../RoomType";
-
+import RoomType from "@docspace/shared/components/room-type";
 import { Base } from "@docspace/shared/themes";
 import { RoomsTypeValues } from "@docspace/shared/utils/common";
