+++ packages/client/src/components/dialogs/CreateEditRoomDialog/sub-components/SetRoomParams.js
@@ -31,8 +31,7 @@ import { inject, observer } from "mobx-react";
 
 import RoomTypeDropdown from "./RoomTypeDropdown";
 import TagInput from "./TagInput";
-import RoomType from "./RoomType";
-
+import RoomType from "@docspace/shared/components/room-type";
 import PermanentSettings from "./PermanentSettings";
 import InputParam from "./Params/InputParam";
 import ThirdPartyStorage from "./ThirdPartyStorage";
