+++ packages/shared/components/room-type/RoomType.types.ts
@@ -1,0 +1,13 @@
+import React from "react";
+import { RoomsType } from "../../enums/index";
+
+export type RoomTypeProps = {
+  roomType: RoomsType;
+  isOpen: boolean;
+
+  type?: "listItem" | "dropdownButton" | "dropdownItem";
+  id?: string;
+  selectedId: string | number;
+
+  onClick: (e: React.MouseEvent<HTMLDivElement>) => void;
+};
